[Decomp][Player] Cohesion split: extract spell-learning methods to PlayerLearn.cpp
