+++ src/game/Object/Player.cpp
@@ -7737,229 +7737,6 @@ void Player::ApplyEquipCooldown(Item* pItem)
 }
 
 /**
- * @brief Resets the player's learned spells and restores default and quest rewards.
- */
-void Player::resetSpells()
-{
-    // not need after this call
-    if (HasAtLoginFlag(AT_LOGIN_RESET_SPELLS))
-    {
-        RemoveAtLoginFlag(AT_LOGIN_RESET_SPELLS, true);
-    }
-
-    // make full copy of map (spells removed and marked as deleted at another spell remove
-    // and we can't use original map for safe iterative with visit each spell at loop end
-    PlayerSpellMap smap = GetSpellMap();
-
-    for (PlayerSpellMap::const_iterator iter = smap.begin(); iter != smap.end(); ++iter)
-    {
-        removeSpell(iter->first, false, false);              // only iter->first can be accessed, object by iter->second can be deleted already
-    }
-
-    learnDefaultSpells();
-    learnQuestRewardedSpells();
-}
-
-/**
- * @brief Teaches the player's default race and class spells.
- */
-void Player::learnDefaultSpells()
-{
-    // learn default race/class spells
-    PlayerInfo const* info = sObjectMgr.GetPlayerInfo(getRace(), getClass());
-    for (PlayerCreateInfoSpells::const_iterator itr = info->spell.begin(); itr != info->spell.end(); ++itr)
-    {
-        uint32 tspell = *itr;
-        DEBUG_LOG("PLAYER (Class: %u Race: %u): Adding initial spell, id = %u", uint32(getClass()), uint32(getRace()), tspell);
-        if (!IsInWorld())                                   // will send in INITIAL_SPELLS in list anyway at map add
-        {
-            addSpell(tspell, true, true, true, false);
-        }
-        else                                                // but send in normal spell in game learn case
-        {
-            learnSpell(tspell, true);
-        }
-    }
-}
-
-/**
- * @brief Teaches a quest reward spell if the quest grants a learnable spell.
- *
- * @param quest The rewarded quest to inspect.
- */
-void Player::learnQuestRewardedSpells(Quest const* quest)
-{
-    uint32 spell_id = quest->GetRewSpellCast();
-
-    // skip quests without rewarded spell
-    if (!spell_id)
-    {
-        return;
-    }
-
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id);
-    if (!spellInfo)
-    {
-        return;
-    }
-
-    // check learned spells state
-    bool found = false;
-    for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-    {
-        if (SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(SpellEffectIndex(i)))
-        {
-            if (spellEffect->Effect == SPELL_EFFECT_LEARN_SPELL && !HasSpell(spellEffect->EffectTriggerSpell))
-            {
-                found = true;
-                break;
-            }
-        }
-    }
-
-    // skip quests with not teaching spell or already known spell
-    if (!found)
-    {
-        return;
-    }
-
-    // prevent learn non first rank unknown profession and second specialization for same profession)
-    SpellEffectEntry const* spellEffect = spellInfo->GetSpellEffect(EFFECT_INDEX_0);
-    uint32 learned_0 = spellEffect ? spellEffect->EffectTriggerSpell : 0;
-
-    if (sSpellMgr.GetSpellRank(learned_0) > 1 && !HasSpell(learned_0))
-    {
-        // not have first rank learned (unlearned prof?)
-        uint32 first_spell = sSpellMgr.GetFirstSpellInChain(learned_0);
-        if (!HasSpell(first_spell))
-        {
-            return;
-        }
-
-        SpellEntry const* learnedInfo = sSpellStore.LookupEntry(learned_0);
-        if (!learnedInfo)
-        {
-            return;
-        }
-
-        // specialization
-        SpellEffectEntry const* learnedSpellEffect0 = learnedInfo->GetSpellEffect(EFFECT_INDEX_0);
-        SpellEffectEntry const* learnedSpellEffect1 = learnedInfo->GetSpellEffect(EFFECT_INDEX_1);
-        if (learnedSpellEffect0 && learnedSpellEffect0->Effect == SPELL_EFFECT_TRADE_SKILL && learnedSpellEffect1 && learnedSpellEffect1->Effect == 0)
-        {
-            // search other specialization for same prof
-            for (PlayerSpellMap::const_iterator itr = m_spells.begin(); itr != m_spells.end(); ++itr)
-            {
-                if (itr->second.state == PLAYERSPELL_REMOVED || itr->first == learned_0)
-                {
-                    continue;
-                }
-
-                SpellEntry const* itrInfo = sSpellStore.LookupEntry(itr->first);
-                if (!itrInfo)
-                {
-                    return;
-                }
-
-                // compare only specializations
-                SpellEffectEntry const* itrSpellEffect0 = learnedInfo->GetSpellEffect(EFFECT_INDEX_0);
-                SpellEffectEntry const* itrSpellEffect1 = learnedInfo->GetSpellEffect(EFFECT_INDEX_1);
-                if ((itrSpellEffect0 && itrSpellEffect0->Effect != SPELL_EFFECT_TRADE_SKILL) || (itrSpellEffect1 && itrSpellEffect1->Effect != 0))
-                {
-                    continue;
-                }
-
-                // compare same chain spells
-                if (sSpellMgr.GetFirstSpellInChain(itr->first) != first_spell)
-                {
-                    continue;
-                }
-
-                // now we have 2 specialization, learn possible only if found is lesser specialization rank
-                if (!sSpellMgr.IsHighRankOfSpell(learned_0, itr->first))
-                {
-                    return;
-                }
-            }
-        }
-    }
-
-    CastSpell(this, spell_id, true);
-}
-
-/**
- * @brief Reapplies all quest reward spells from previously rewarded quests.
- */
-void Player::learnQuestRewardedSpells()
-{
-    // learn spells received from quest completing
-    for (QuestStatusMap::const_iterator itr = mQuestStatus.begin(); itr != mQuestStatus.end(); ++itr)
-    {
-        // skip no rewarded quests
-        if (!itr->second.m_rewarded)
-        {
-            continue;
-        }
-
-        Quest const* quest = sObjectMgr.GetQuestTemplate(itr->first);
-        if (!quest)
-        {
-            continue;
-        }
-
-        learnQuestRewardedSpells(quest);
-    }
-}
-
-/**
- * @brief Learns or removes spells unlocked by a profession or skill value.
- *
- * @param skill_id The skill line identifier.
- * @param skill_value The current skill value.
- */
-void Player::learnSkillRewardedSpells(uint32 skill_id, uint32 skill_value)
-{
-    uint32 raceMask  = getRaceMask();
-    uint32 classMask = getClassMask();
-    for (uint32 j = 0; j < sSkillLineAbilityStore.GetNumRows(); ++j)
-    {
-        SkillLineAbilityEntry const* pAbility = sSkillLineAbilityStore.LookupEntry(j);
-        if (!pAbility || pAbility->skillId != skill_id || pAbility->learnOnGetSkill != ABILITY_LEARNED_ON_GET_PROFESSION_SKILL)
-        {
-            continue;
-        }
-        // Check race if set
-        if (pAbility->racemask && !(pAbility->racemask & raceMask))
-        {
-            continue;
-        }
-        // Check class if set
-        if (pAbility->classmask && !(pAbility->classmask & classMask))
-        {
-            continue;
-        }
-
-        if (sSpellStore.LookupEntry(pAbility->spellId))
-        {
-            // need unlearn spell
-            if (skill_value < pAbility->req_skill_value)
-            {
-                removeSpell(pAbility->spellId);
-            }
-            // need learn
-            else if (!IsInWorld())
-            {
-                addSpell(pAbility->spellId, true, true, true, false);
-            }
-            else
-            {
-                learnSpell(pAbility->spellId, true);
-            }
-        }
-    }
-}
-
-/**
  * @brief Sends visible aura duration updates for a target to the player.
  *
  * @param target The unit whose aura durations should be sent.
